MetadataGenerator.Api/pyproject.toml
[project]
name = "metadata-generator-api"
version = "0.1.0"
description = "Backend API for AI-powered metadata extraction from images and audio files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.18",
    "azure-ai-contentunderstanding>=1.0.0b1",
    "pillow>=12.1.1",
    "mutagen>=1.47.0",
    "httpx>=0.28.0",
    "aiohttp>=3.13.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "RUF",  # ruff-specific rules
]
ignore = [
    "S104",  # Possible binding to all interfaces (needed for dev server)
    "B008",  # Depends() in function defaults is a FastAPI convention
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # assert is expected in test files

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 85
show_missing = true
